Fix 64-bit warning on format string

+++ RZUtils/Components/RZAnimatedCountingLabel/RZAnimatedCountingLabel.m
@@ -141,7 +141,7 @@
     }
     else
     {
-        valueString = [NSString stringWithFormat:@"%d", self.currentNumber.integerValue];
+        valueString = [NSString stringWithFormat:@"%ld", (long)self.currentNumber.integerValue];
     }
     
     self.text = valueString;
